+++ tpm/platform/src/RunCommand.c
@@ -72,7 +72,7 @@ _plat__RunCommand(
     unsigned char   *request,       // IN: command buffer
     uint32_t        *responseSize,  // IN/OUT: response buffer size
     unsigned char   **response,      // IN/OUT: response buffer
-    uint32_t         contextId
+    __uint128_t      contextId
     )
 {
     // setjmp(s_jumpBuffer);
@@ -96,7 +96,7 @@ _plat__Fail(
 
 LIB_EXPORT int
 _plat__TPM_Terminate(
-    uint32_t        contextId
+    __uint128_t        contextId
 )
 {
     SwitchTpmContext(contextId);
@@ -126,7 +126,7 @@ _plat__ReadTsc(
 
 LIB_EXPORT int
 _plat__TPM_Initialize(
-    uint32_t        contextId,
+    __uint128_t     contextId,
     int             firstTime       // IN: indicates if this is the first call from
                                     //     main()
 )
